Allow overriding the time-of-day Fitbit returns for weighings
Since it only stores one reading per day, and forgets the time,
it always returns 23:59:59, which looks kinda weird in Fit.
For one thing, for today it's usually in the future.
If you weigh yourself in the morning, this'll be off by most of a
day, every day, which I found irritating.